Refactor code for norminette
Improved the structure and formatting of the code in various files such as 'philosophers.c', 'manage_dinner.c', 'common.c', 'set_table.c' in order to enhance clarity and maintainability. The changes include better line breaks, well-organized block structures, and improved arrangement of assignments. The goal is to make the code easier to read, understand, debug, and update in future while the logic and function of the code remain the same.