feat(philo): refactor fork and meal management
Mutex locks were added for accessing meal count and for the right fork in the `take_forks` function. TODO notes for checking if `time_to_eat` is set and implementing meal mutex when eating are also inserted. Additionally, the increase in `fed_count` was updated to only occur when there are no meals.